// cli/src/types/openapi-schemas.ts
export const WorkspaceSummaryResponse = z
  .object({
    current: z.boolean(),
    id: z.string(),
    name: z.string(),
    role: z.string(),
    status: z.string(),
  })

export const WorkspaceListResponse = z
  .object({ workspaces: z.array(WorkspaceSummaryResponse) })
// ...
export type WorkspaceListResponseType = z.infer<typeof WorkspaceListResponse>